+++ test/gurka/test_hard_exclusions_highways.cc
@@ -20,7 +20,7 @@ const std::vector<std::string> kExclusionParameters = {"exclude_highways"};
 constexpr double grid_size_meters = 100.;
 
 const std::string ascii_map = R"(
-  E----F----G----H----I----A----J----K----L----O----P----Q----R
+  E====F----G----H----I----A----J====K----L====O====P----Q----R
                                 |    |         |    |
                                 |    |         |    |
                                 |    |         |    |
@@ -31,7 +31,7 @@ const std::string ascii_map = R"(
 const gurka::ways ways = {
     {"EF", {{"highway", "motorway"}}},    {"FG", {{"highway", "residential"}}},
     {"GH", {{"highway", "residential"}}}, {"HI", {{"highway", "residential"}}},
-    {"IA", {{"highway", "residential"}}}, {"IJ", {{"highway", "residential"}}},
+    {"IA", {{"highway", "residential"}}}, {"AJ", {{"highway", "residential"}}},
     {"JK", {{"highway", "motorway"}}},    {"KL", {{"highway", "residential"}}},
     {"JM", {{"highway", "residential"}}}, {"MN", {{"highway", "residential"}}},
     {"NK", {{"highway", "residential"}}}, {"LO", {{"highway", "motorway"}}},
@@ -50,9 +50,12 @@ void check_result(const std::string& exclude_parameter_value,
 
   const std::string& costing = props[0];
   const std::string& exclude_parameter = props[1];
+  const std::string& time_type = props[2];
   const auto result = gurka::do_action(valhalla::Options::route, map, waypoints, costing,
                                        {{"/costing_options/" + costing + "/" + exclude_parameter,
-                                         exclude_parameter_value}});
+                                         exclude_parameter_value},
+                                        {"/date_time/type", time_type},
+                                        {"/date_time/value", "2025-07-21T12:00"}});
   gurka::assert::raw::expect_path(result, expected_names);
 }
 
@@ -75,7 +78,7 @@ gurka::map ExclusionTestExcludeHighways::map = {};
 gurka::map ExclusionTestExcludeHighways::mapNotAllowed = {};
 
 TEST_P(ExclusionTestExcludeHighways, ExcludeHighways) {
-  check_result("1", {"I", "L"}, {"IJ", "JM", "MN", "NK", "KL"}, map, GetParam());
+  check_result("1", {"I", "L"}, {"IA", "AJ", "JM", "MN", "NK", "KL"}, map, GetParam());
 }
 
 TEST_P(ExclusionTestExcludeHighways, InTheBeginningWithNoExit) {
@@ -91,11 +94,17 @@ TEST_P(ExclusionTestExcludeHighways, InTheBeginningWithPossibleExit) {
 
 TEST_P(ExclusionTestExcludeHighways, InTheBeginningNotAllowed) {
   check_result("0", {"E", "H"}, {"EF", "FG", "GH"}, mapNotAllowed, GetParam());
+  check_result("1", {"E", "H"}, {"EF", "FG", "GH"}, mapNotAllowed, GetParam());
+}
+
+TEST_P(ExclusionTestExcludeHighways, InTheEndNotAllowed) {
+  check_result("0", {"H", "E"}, {"GH", "FG", "EF"}, mapNotAllowed, GetParam());
   try {
-    check_result("1", {"E", "H"}, {"EF", "FG", "GH"}, mapNotAllowed, GetParam());
-  } catch (const valhalla_exception_t& err) { EXPECT_EQ(err.code, 145); } catch (...) {
+    check_result("1", {"H", "E"}, {"EF", "FG", "GH"}, mapNotAllowed, GetParam());
+    FAIL() << "Missing exception.";
+  } catch (const valhalla_exception_t& err) { EXPECT_EQ(err.code, 442); } catch (...) {
     FAIL() << "Expected valhalla_exception_t.";
-  };
+  }
 }
 
 INSTANTIATE_TEST_SUITE_P(ExcludeHighwaysTests,
@@ -104,7 +113,9 @@ INSTANTIATE_TEST_SUITE_P(ExcludeHighwaysTests,
                            std::vector<std::vector<std::string>> values;
                            for (const auto& costing : kCostingModelsExcludeHighways) {
                              for (const auto& param : kExclusionParameters) {
-                               values.push_back({costing, param});
+                               for (uint32_t time_type = 0; time_type < 4; time_type++) {
+                                 values.push_back({costing, param, std::to_string(time_type)});
+                               }
                              }
                            }
                            return values;
@@ -123,7 +134,7 @@ protected:
 gurka::map ExclusionTestNoHardExcludeHighways::map = {};
 
 TEST_P(ExclusionTestNoHardExcludeHighways, NoHardExcludeHighways) {
-  check_result("0", {"I", "L"}, {"IJ", "JK", "KL"}, map, GetParam());
+  check_result("0", {"I", "L"}, {"IA", "AJ", "JK", "KL"}, map, GetParam());
 }
 
 INSTANTIATE_TEST_SUITE_P(NoHardExcludeHighwaysTests,
@@ -132,7 +143,9 @@ INSTANTIATE_TEST_SUITE_P(NoHardExcludeHighwaysTests,
                            std::vector<std::vector<std::string>> values;
                            for (const auto& costing : kCostingModelsNoHardExcludeSetA) {
                              for (const auto& param : kExclusionParameters) {
-                               values.push_back({costing, param});
+                               for (uint32_t time_type = 0; time_type < 4; time_type++) {
+                                 values.push_back({costing, param, std::to_string(time_type)});
+                               }
                              }
                            }
                            return values;
@@ -152,7 +165,7 @@ protected:
 gurka::map ExclusionTestNoHardExcludeHighwaysForOtherModes::map = {};
 
 TEST_P(ExclusionTestNoHardExcludeHighwaysForOtherModes, NoHardExcludeHighwaysForOtherModes) {
-  check_result("0", {"I", "L"}, {"IJ", "JM", "MN", "NK", "KL"}, map, GetParam());
+  check_result("0", {"I", "L"}, {"IA", "AJ", "JM", "MN", "NK", "KL"}, map, GetParam());
 }
 
 INSTANTIATE_TEST_SUITE_P(NoHardExcludeHighwaysForOtherModesTests,
@@ -161,7 +174,9 @@ INSTANTIATE_TEST_SUITE_P(NoHardExcludeHighwaysForOtherModesTests,
                            std::vector<std::vector<std::string>> values;
                            for (const auto& costing : kCostingModelsNoHardExcludeSetB) {
                              for (const auto& param : kExclusionParameters) {
-                               values.push_back({costing, param});
+                               for (uint32_t time_type = 0; time_type < 4; time_type++) {
+                                 values.push_back({costing, param, std::to_string(time_type)});
+                               }
                              }
                            }
                            return values;